refactoring to be release {{type}} instead of release-{{type}}
