+++ internal/botcommander/commands/init.go
@@ -24,8 +24,7 @@ func NewFactory() Factory {
 			ShowCmdName:             NewShowCommand,
 			SetCmdName:              NewSetCommand,
 			DeleteCmdName:           NewDeleteCommand,
-			ReleaseArtifactCmdName:  NewReleaseArtifactCommand,
-			ReleaseNamespaceCmdName: NewReleaseNamespaceCommand,
+			ReleaseCmdName:          NewReleaseCommand,
 			RestartCmdName:          NewRestartCommand,
 			RunCmdName:              NewRunCommand,
 			AuthCmdName:             NewAuthCommand,
